--- src/renderer/src/components/github/github-issue-comment-helpers.ts
@@ -1,3 +1,0 @@
-export function githubAvatarUrl(login: string): string {
-  return `https://github.com/${encodeURIComponent(login)}.png?size=64`
-}
