+++ src/renderer/src/components/github/github-user-avatar.tsx
@@ -1,6 +1,9 @@
 import { useState } from 'react'
 import { cn } from '@/lib/utils'
-import { githubAvatarUrl } from '@/components/github/github-issue-comment-helpers'
+
+export function githubAvatarUrl(login: string): string {
+  return `https://github.com/${encodeURIComponent(login)}.png?size=64`
+}
 
 /**
  * Build a 1-2 character initials placeholder from a display name or login,
